I have done a little refactor with matchOrch and checkLetters. I have created a matchFiveLetterWords.js with its 4 counterpart
