+++ components/checkWords/checkLettersUsed.js
@@ -1,12 +1,22 @@
 
 
-module.exports = function checkLetersUsed(matchedWord1, matchedWord2, matchedWord3, matchedWord4, sentLetters) {
+module.exports = function checkLetersUsed(sentLetters, matchedWord1, matchedWord2, matchedWord3, matchedWord4, matchedWord5) {
     try {
-        let wordsAdded = matchedWord1 + matchedWord2 + matchedWord3 + matchedWord4
-        if (wordsAdded.split('').sort().join() == sentLetters.split('').sort().join()) {
-            return true
+        let wordsAdded
+        if (matchedWord5 == undefined) {
+            wordsAdded = matchedWord1 + matchedWord2 + matchedWord3 + matchedWord4
+            if (wordsAdded.split('').sort().join() == sentLetters.split('').sort().join()) {
+                return true
+            } else {
+                return false
+            }
         } else {
-            return false
+            wordsAdded = matchedWord1 + matchedWord2 + matchedWord3 + matchedWord4 + matchedWord5
+            if (wordsAdded.split('').sort().join() == sentLetters.split('').sort().join()) {
+                return true
+            } else {
+                return false
+            }
         }
     } catch (error) {
         throw error
